chore(rust): split the geometry encoder into encode01 and streams
