+++ rust/mlt-core/src/encoder/geometry/mod.rs
@@ -1,6 +1,7 @@
-pub(crate) mod encode;
+mod encode01;
 mod geotype;
 mod model;
+mod streams;
 #[cfg(test)]
 mod tests;
